migrate Index page to ts
